feat: add development configuration file for Odoo
Add a new configuration file `odoo-dev.conf` to the `config` directory. This file contains the following options:
- `admin_passwd`: Set the admin password to "IciMettreMotDePasseSecuriser"
- `addons_path`: Set the addons path to "/mnt/extra-addons/addons/crm,/mnt/extra-addons/addons/custom,/mnt/extra-addons/dev-addons/web"

The purpose of this change is to provide a separate configuration file specifically for development purposes in Odoo.

fix: update addons_path in production configuration file

In the `odoo.conf` file located in the `config` directory, update the `addons_path` option to "/mnt/extra-addons/addons/crm,/mnt/extra-addons/addons/custom". This change ensures that the correct addons path is used in the production environment.

fix: update docker-compose.dev.yml

In the `docker-compose.dev.yml` file, make the following changes:
- Update the image version to "odoo:16.0"
- Update the container name to "odoo-web-dev"
- Add a network configuration for the "odoo-dev-network"
- Update the command to include additional options:
  - `--config=/etc/odoo/odoo-dev.conf`: Specify the configuration file to use
- Add a volume configuration to mount the `config` directory as `/etc/odoo`

The purpose of these changes is to ensure that the development environment is properly configured and uses the correct Odoo image, container name, network, and configuration file.

feat: add production docker-compose file

Add a new file `docker-compose.prod.yml` to define the production environment for Odoo. This file includes the following services:
- `web`: Odoo web container
  - Image: odoo:16.0
  - Container name: odoo-web
  - Network: odoo-network
  - Ports: 8069
  - Command: `odoo --config=/etc/odoo/odoo.conf`
  - Volumes: Mount `addons` and `config` directories, as well as `odoo-data` for data persistence
- `db`: Odoo database container
  - Image: postgres:16
  - Container name: odoo-db
  - Network: odoo-network
  - Volumes: Mount `postgres-data` for data persistence